Update weapon_physcannon.cpp
Added check for attached object. It's fixes elements closing when player take an object and sharply trying to drop it and also pull's sound;
The delay for the secondary attack has been removed, this makes it easier for you to catch fast moving and small objects;
Some checks are replaced with a single function CanPickupObject;
Changed pull force calculations to keep the same force after removing the delay of secondary attack;
Physcannon can only pull objects with physics movetype. This saves you from unpleasant moments, for example, when trying to pick up an object on a moving platform.